feat(sale): atualiza comportamento de venda e fluxo de caixa
- Atualiza metodo create para apenas abrir uma venda vazia.
- Implementa addItemByBarcode para adicionar os itens da venda (balança e prateleira).
- Implementa completeSale para fechar a venda, calculando valor total e método de pagamento.
- Implementa principio ACID na baixa de estoque.
- Atualiza, adiciona e corrige testes unitários do módulo."

+++ src/modules/sale/dto/sale-item.dto.ts
@@ -1,9 +1,9 @@
-import { Field, ID, Int, ObjectType } from '@nestjs/graphql';
+import { Field, Float, ID, Int, ObjectType } from '@nestjs/graphql';
 import { BaseDTO } from 'src/modules/bases/dto/base.dto';
 
 @ObjectType('SaleItem')
 export class SaleItemDTO extends BaseDTO {
-  @Field(() => Int)
+  @Field(() => Float)
   quantity!: number;
 
   @Field(() => Int)
